refactor: rename github_api => GithubApi
