+++ src/App.tsx
@@ -8,7 +8,7 @@ import {
   DefaultGitHubAPI,
   GithubErrorResponse,
   GithubResponse,
-} from './utils/api/github_api';
+} from './utils/api/GithubApi';
 import Results from './components/Result';
 import ErrorBoundary from './ErrorBoundary';
 import ErrorThrower from './ErrorThrower';
